Add more methods to ReportingService
