+++ src/crispin/IReportingService.cs
@@ -3,5 +3,8 @@
     public interface IReportingService
     {
         void OpenAsPdf(string xrpt, string reportName);
+        byte[] ConvertToPdf(string xrpt, string reportName);
+        string ConvertToCsv(string xrpt);
+        string ConvertToHtml(string xrpt);
     }
 }
